Changed the folder structure to original
